api/requirements.txt
Flask==2.3.2
jsonschema==4.17.3
pymongo==4.3.3
celery==5.3.0
waitress==2.1.2
PyJWT==2.7.0